requirements.txt
django==5.2.11
psycopg2-binary==2.9.11
pillow==12.1.8
python-dotenv==1.2.1
djangorestframework==3.14.0
django-filter==24.1
django-cors-headers==4.3.1
djazzmin==3.0.7
daphne==4.1.2
stripe==10.8.1
celery==5.3.6
redis==5.0.1
pip-tools==7.4.0